requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
pyrsistent==0.18.0
pytz==2021.3
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.6.0
